// frontend/src/app/shared/models/feed-post.ts
export class FeedPost {
  // tslint:disable-next-line: variable-name
  _id: string;
  author: string;
  creationDate: Date;
  text: string;
  projectId: string;
  imgUrl: string;
  likes: [];
  projectStart?: boolean;


  // toggleLike(userId: string) {}

  // addImage(imgUrl: string) {}

  // createPost(feedPost: FeedPost) {}
}